Set a default value of 'none' to instance attribute in alerts
Sometimes, an alert is generated without an instance.
For example if an alert is triggered when none of the monitored
instances has a particular property, such as the case when a vital
container is *not* running on any instance.

In this case, it makes sense to set the value of the `instance`
property to `none`, since it did not originate from a particular
intance.

This PR changes `intance` to an optional field, and sets a default value
to `none`.

Closes #32